pi/requirements.txt
absl-py==0.8.0
Adafruit-Blinka==3.0.0
adafruit-circuitpython-busdevice==4.0.0
adafruit-circuitpython-rfm9x==1.1.7
Adafruit-PlatformDetect==1.3.4
Adafruit-PureIO==0.2.3
astor==0.8.0
audioread==2.1.8
cffi==1.12.3
colorzero==1.1
cycler==0.10.0
decorator==4.4.0
gast==0.3.1
google-pasta==0.1.7
gpiozero==1.5.1
grpcio==1.23.0
h5py==2.10.0
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.4.2
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
pkg-resources==0.0.0
protobuf==3.9.1
PyAudio==0.2.11
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
pyusb==1.0.2
PyYAML==5.1.2
resampy==0.1.4rc0
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
scikit-learn==0.21.3
scipy==1.2.2
six==1.12.0
sounddevice==0.3.13
SoundFile==0.10.2
spidev==3.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.14.0
termcolor==1.1.0
voice-engine==0.1.3
webrtcvad==2.0.10
Werkzeug==0.15.6
wrapt==1.11.2